HDMI and VGA monitors are swapped
Due to a physical change in my setup, I had to swap the VGA and HDMI
cable connections, making the monitors set up the other way around.

To make that match my settings I changed all the values involving the
HDMI and VGA outputs so that everything stays the same software wise.