Sublime Text 3 keyboard shortcut added
I mainly use emacs for any text editing but I'm giving Sublime Text 3 a
go for some things. To make it easier to open Sublime whenever I want, I
set a keyboard shortcut for it too. It's the same key as emacs but with
an added Shift (emacs stills my primary editor hence it keeps the easier
calling method)