i3-wm changes for new install
I changed my main machine and with that some settings from my old laptop
install were not entirely compatible.

First and foremost the monitor setup changed slightly due to the
connection type and position of each screen, with that the settings
involving binding applications to desktops and desktop positions had to
be updated.

There was also an update on one of my scripts that loaded the
arrangement of both screens and set up my desktop wallpaper. That script
was removed and the settings are now applied directly throught i3's
config file (see Dual Monitors and Set Wallpaper).

Following up there was a change in my zsh's aliases, which allowed me to
revert emacs daemon init to just emacs, rather than emacs-24.4.

And lastly, for i3, I also changed my main music player to Spotify (for
convenience and interoperability with my dayjob setup). This forced me
to change the shortcuts that controlled music playing.

The next set of changes involve the i3 status bar, a.k.a. i3pystatus.

Because I now use a full keyboard a numpad switch was added. There were
also a few additions related to CPU and RAM memory displays, due to
updates in i3pystatus. The i3status.py file is pretty well documented so
I don't believe there's a need to add verbose documentation in here. You
should be able to just follow that file through along with the
documentation link noted in the beginning of the file.